+++ src/voice.ts
@@ -24,6 +24,7 @@ export interface IMusic {
     url: string;
     detail: VideoDetails;
     rawmeta: videoInfo;
+    requested_by: string;
 }
 
 // @ts-ignore
@@ -36,21 +37,25 @@ interface MusicState {
     is_looping: boolean;
     is_playing: boolean;
     channel_id: string | null;
+    playing_since: number;
 }
 
-const defaultMusicState: MusicState = {
-    music_queue: [],
-    now_playing: undefined,
-    audio_player: null,
-    is_looping: false,
-    is_playing: false,
-    channel_id: null,
-};
+function defaultMusicState() {
+    return {
+        music_queue: [],
+        now_playing: undefined,
+        audio_player: null,
+        is_looping: false,
+        is_playing: false,
+        channel_id: null,
+        playing_since: 0,
+    };
+}
 
 export const musicStates: { [guildId: string]: MusicState } = {};
 
 export function getState(guildId: string) {
-    return (musicStates[guildId] ??= defaultMusicState);
+    return (musicStates[guildId] ??= defaultMusicState());
 }
 
 export namespace VoiceHelper {
@@ -75,31 +80,43 @@ export namespace YoutubeHelper {
 }
 
 export namespace Voice {
+    export enum JoinFailureReason {
+        Success,
+        AlreadyConnected,
+        NoChannel,
+        NotJoinable,
+        Other,
+    }
+
     /**
      * Joins to the channel if not already in one.
-     * @returns `false` if no changes, `true` if new channel is joined
      */
-    export async function joinFromContext(ctx: Context) {
+    export async function joinFromContext(
+        ctx: Context,
+        force = false
+    ): Promise<JoinFailureReason> {
         const connection = getVoiceConnection(ctx.guildId!);
 
-        if (connection?.state.status == VoiceConnectionStatus.Ready) {
-            return false;
+        if (connection?.state.status == VoiceConnectionStatus.Ready && !force) {
+            return JoinFailureReason.AlreadyConnected;
         }
 
         const voiceChannel = (ctx.member as GuildMember | undefined)?.voice
             .channel as VoiceChannel | undefined;
 
-        if (!voiceChannel) return false;
+        if (!voiceChannel) return JoinFailureReason.NoChannel;
 
-        const guild = ctx.client.guilds.cache.get(ctx.guildId!);
+        if (!voiceChannel.joinable) return JoinFailureReason.NotJoinable;
 
-        if (!guild?.available) return false;
-
-        await Voice.joinVoiceChannel(voiceChannel);
-
-        return true;
+        return (await Voice.joinVoiceChannel(voiceChannel))
+            ? JoinFailureReason.Success
+            : JoinFailureReason.Other;
     }
 
+    /**
+     * Joins voice channel
+     * @returns Connection
+     */
     export async function joinVoiceChannel(
         channel: VoiceChannel,
         onDisconnect?: () => Awaitable<void>
@@ -136,6 +153,7 @@ export namespace Voice {
 
         try {
             await entersState(connection, VoiceConnectionStatus.Ready, 5_000);
+            getState(channel.guildId).channel_id = channel.id;
             return connection;
         } catch (err) {
             return undefined;
@@ -144,9 +162,13 @@ export namespace Voice {
 
     /**
      * Add music to queue and play it if not playing
-     * @returns Meta Info of the Video
+     * @returns Meta Info of the Video or string indicating failure reason
      */
-    export async function addMusicToQueue(guildId: string, url: string) {
+    export async function addMusicToQueue(
+        guildId: string,
+        url: string,
+        requester: string
+    ) {
         if (!ytdl.validateURL(url)) {
             url = (await YoutubeHelper.searchVideo(url))[0]?.link ?? "";
             if (!url) return "No results found";
@@ -156,7 +178,12 @@ export namespace Voice {
         const detail = meta.player_response.videoDetails;
 
         const state = getState(guildId);
-        state.music_queue.push({ url, detail, rawmeta: meta });
+        state.music_queue.push({
+            url,
+            detail,
+            rawmeta: meta,
+            requested_by: requester,
+        });
 
         if (!state.is_playing) playNextMusicInQueue(guildId);
 
@@ -165,8 +192,8 @@ export namespace Voice {
 
     /**
      * @param guildId
-     * @returns true if music finished successfully,
-     * false immediately if no connection found or later when error occured
+     * @returns `true` if music finished successfully,
+     * `false` early if no connection found or later when error occured
      */
     export function playNextMusicInQueue(guildId: string) {
         const state = getState(guildId);
@@ -178,6 +205,7 @@ export namespace Voice {
         if (state.music_queue.length < 1) {
             state.is_playing = false;
             VoiceHelper.forceDestroyConnection(getVoiceConnection(guildId));
+            state.channel_id = null;
             return;
         }
 
@@ -202,6 +230,7 @@ export namespace Voice {
         audioPlayer.play(resource);
 
         state.is_playing = true;
+        state.playing_since = new Date().getTime();
 
         return new Promise<boolean>((resolve, reject) => {
             audioPlayer.on(AudioPlayerStatus.Idle, () => {