+++ CHANGELOG.md
@@ -1,3 +1,26 @@
 # Changelog
 
 All notables change for @leomotors/music-bot will be noted here
+
+## [0.9.0] - 2022-06-22
+
+- Update to `cocoa-discord-utils` 1.9 which supports `discord.js` 13.8
+
+## [0.7.0] - 2022-06-11
+
+- Fixed bugs in /now
+- Update /now
+
+## [0.5.0] - 2022-06-09
+
+- Added /now
+
+## [0.4.0] - 2022-05-06
+
+- The bot now tell user if it is attempting to joining empty or not joinable functions
+
+- Add /rejoin to force bot to join the voice channel
+
+## [0.3.0] - 2022-05-03
+
+- Fixed an issue where all server share the same music state (Reference and Value Stuff, again)